Move more container related components into src/components/container
