+++ app/packs/src/components/container/AnalysisEditor.js
